// frontend/libs/iam/data-access/src/lib/generated/models/access-resource-dto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ResourceCapabilityDto } from './resource-capability-dto'

export interface AccessResourceDto {
	id: string
	key: string
	name: string
	capabilities: ResourceCapabilityDto[]
	children: AccessResourceDto[]
}