// frontend/libs/iam/data-access/src/lib/generated/models/resource-capability-dto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AccessScope } from './access-scope'
import type { PermissionAction } from './permission-action'

export interface ResourceCapabilityDto {
	action: PermissionAction
	availableScopes: AccessScope[]
}